pyproject.toml
[project]
name = "llm-providers-explorer"
version = "0.1.0"
description = "Some utils for open llms providers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.26.0",
    "groq>=0.30.0",
    "mistralai>=1.9.2",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "together>=1.5.21",
]
